+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * Tooltip: quests completed by other chars were not indicated as active.
 * Tooltip: zone story chapter quests were only shown when eg. a chapter wasn't indicated as active.
+* Settings: some achievement entries in the zone view tooltip for optional achievements could not be hidden when there were more than one.
 * Settings: some zone icons in continent view for optional achievements could not be hidden.
 
 ## [0.7.1-beta+100206] - 2024-03-22